# Replace a call to PyTuple_New with _PyTuple_FromArraySteal