@mdboom 2022-09-02
This is a great small change, especially if it ends up providing most of the benefit of #96446.

I'm running this PR on the Faster CPython team's standard benchmarking machine as well to get another data point (though I may not be able to report back until after the long weekend).