@mdboom 2022-09-02
I measured only a 0.07% speedup on the pyperformance and pyston macrobenchmarks. I'm going to run this (and the baseline) again, because it's surprising it's so little.  I don't want to imply that's an argument against this PR (and it's not really up to me anyway).  Reproducibility of benchmarks in general, especially when the margins are so small, is an ongoing challenge we're all struggling with.