@kmod 2022-09-03
Yeah I think that even with some changes to make benchmarking more stable (bolt + [longer bolt task](https://github.com/kmod/cpython/tree/longer_pgo_task)) the systematic errors are around this level so I think it's hard to rely on the benchmarking numbers too much.