@methane 2022-09-14
This improvement is very small to users. I think we don't need news file for this.

```
$ ./python-nonzero2 -m pyperf timeit --duplicate 10 --compare-to ./python --python-names master:patched -s 'x=42' -- '(1,2,x)'
master: ..................... 29.1 ns +- 0.1 ns
patched: ..................... 29.3 ns +- 0.3 ns

Mean +- std dev: [master] 29.1 ns +- 0.1 ns -> [patched] 29.3 ns +- 0.3 ns: 1.01x slower

$ ./python-nonzero2 -m pyperf timeit --duplicate 10 --compare-to ./python --python-names master:patched -s 'x=42' -- '(1,2,3,4,5,6,7,8,9,x)'
master: ..................... 56.0 ns +- 0.5 ns
patched: ..................... 55.5 ns +- 1.3 ns

Mean +- std dev: [master] 56.0 ns +- 0.5 ns -> [patched] 55.5 ns +- 1.3 ns: 1.01x faster
```